compute: merge materialized_view_v2 into materialized_view
With the v1 sink gone, the sync-operator implementation in
materialized_view_v2 is the only one, so the split into two modules no
longer carries its weight. Fold the persist_sink entry point and the mint,
write, and append operator modules back into materialized_view, next to the
SinkRender impl and shared scaffolding they use, and delete the now-empty
materialized_view_v2 module. The channel-ordering invariant docs move into
the main module-level documentation. Pure code movement; no behavior change.
