+++ src/compute/src/sink.rs
@@ -17,6 +17,5 @@ pub mod correction_v2;
 #[cfg(not(feature = "bench"))]
 mod correction_v2;
 mod materialized_view;
-mod materialized_view_v2;
 mod refresh;
 mod subscribe;
